clippy: Remove unnecessary referencing

+++ core/storage/subjournal.rs
@@ -70,7 +70,7 @@ impl Subjournal {
                     bytes_read == page_size as i32,
                     "bytes_read should be page_size"
                 );
-                finish_read_page(page.get().id as usize, buffer, page.clone());
+                finish_read_page(page.get().id, buffer, page.clone());
             },
         );
         let c = self.file.pread(offset, c)?;
